image_zfs: destroy correct zpool on cleanup
image_zfs uses `zpool create -t` to create an in-core temporary pool
name while the on-disk name will be set to $ZFS_POOL_NAME, the default
value of $ZFS_POOL_NAME is "zroot".

If an error occurs while running image_zfs, the on-disk name is used
when destroying the zpool on cleanup, causing the wrong pool to be
destroyed.

Fix this by storing the temporary pool name in the zroot variable since
image.sh uses the zroot variable to destroy the zpool on cleanup.

Fixes:  73b3a51: image_zfs: Allow same-as-host pool names
Sponsored-by: Klara, Inc.